Online users list: modify the ORDER BY clause in SQL query so that groupid=7 and groupid=8 are sorted after other group IDs (like groupid=15) Currently, it's ORDER BY groupid<7 DESC, lastactivity DESC. This sorts rows where groupid is less than 7 first (descending order of the boolean result of groupid<7).
We want rows where groupid is 7 or 8 to come last. This means rows where groupid is not 7 and not 8 should come first.

We can achieve this by changing groupid<7 DESC to (groupid NOT IN (7, 8)) DESC.

This change should give the desired sorting behavior where groupid=15 (and others not 7 or 8) are prioritized over groupid=7 and groupid=8.