the "Who's Online" list should update every 30 seconds by fetching new data via AJAX.
In `forum_index.php`, the section that calculates online statistics and prepares `$whosonline` (around lines 331-370) needs adjustment. We want it to still calculate `$membercount`, `$guestcount`, and `$invisiblecount` for the summary line, but it no longer needs to build the full `$whosonline` array for the list.

Modify `source/module/forum/forum_ajax.php`. Add a new action to handle fetching and rendering the online list.
template/default/forum/ajax_whosonline_list.htm`: This new template file will be used by the AJAX action to render the list items.

template/default/forum/discuz.htm : This template will be updated to remove the static rendering of the online list and instead use JavaScript to fetch and display it dynamically.