mode: The mode you want to use for the request, e.g., cors, no-cors, same-origin, or navigate. The default is cors.
So there's no need to specify { mode: "cors" } with fetch - that's the default.

the specs says the default is: {mode:"no-cors"}

For example, when a Request object is created using the Request.Request constructor, the value of the mode property for that Request is set to cors.

However, for requests created other than by the Request.Request constructor, no-cors is typically used as the mode; for example, for embedded resources where the request is initiated from markup, unless the crossorigin attribute is present, the request is in most cases made using the no-cors mode — that is, for the <link> or <script> elements (except when used with modules), or <img>, <audio>, <video>, <object>, <embed>, or <iframe> elements.